requirements.txt
Pillow==8.1.2
typing==3.7.4.3